fix: widen Unpack to u32, add error messages and num_regs OOM guard
- Widen Insn::Unpack n from u8 to u32 so tuple unpacking isn't silently
  capped at 255 targets
- Replace dead Reg::MAX / hardcoded 256 overflow checks with
  checked_add for all alloc paths affected by the u8→u32 reg widening
- Add error_msg to every failed=true path that previously produced the
  generic "compilation failed" string (intern overflow, break/continue
  outside loop, unsupported targets, fn_proto limit, frame overflow, etc.)
- Fix alloc_temp error message to say "temporaries" instead of "local
  variables"
- Add MAX_REGS (1<<20) guard in finish() to prevent OOM on pathological
  register counts
- Simplify try_exec_vm_script_with_index: hoist Rc::new outside match
- Update stale comment in try_exec_vm_script (was "Reg type is u8")

+++ src/bytecode.rs
@@ -113,7 +113,7 @@ pub enum Insn {
     /// R[dst] = {R[base]: R[base+1], R[base+2]: R[base+3], ...}  (n key-value pairs)
     BuildDict(Reg, Reg, u8),
     /// R[base..base+n] = iter_values(R[src])
-    Unpack(Reg, Reg, u8),
+    Unpack(Reg, Reg, u32),
     /// iters[slot] = iter_values(R[src])
     GetIter(u8, Reg),
     /// if iters[slot] exhausted: pc += offset; else R[dst] = next(iters[slot])
